fix: job - create job
