+++ src/models/job.ts
@@ -1,40 +1,10 @@
-import {
-  Entity,
-  PrimaryGeneratedColumn,
-  Column,
-  ManyToOne,
-  UpdateDateColumn,
-  CreateDateColumn,
-} from "typeorm";
+import { Entity, Column, ManyToOne } from "typeorm";
 import ExtendedBaseEntity from "./base-entity";
 import { User } from "./user";
-
-export enum SalaryRange {
-  below_30k = "below_30k",
-  k_30k_to_50k = "30k_to_50k",
-  k_50k_to_70k = "50k_to_70k",
-  k_70k_to_100k = "70k_to_100k",
-  k_100k_to_150k = "100k_to_150k",
-  above_150k = "above_150k",
-}
-
-export enum JobType {
-  full_time = "full-time",
-  part_time = "part-time",
-  internship = "internship",
-  contract = "contract",
-}
-
-export enum JobMode {
-  remote = "remote",
-  onsite = "onsite",
-}
+import { JobMode, JobType, SalaryRange } from "../types";
 
 @Entity()
 export class Job extends ExtendedBaseEntity {
-  @PrimaryGeneratedColumn("uuid")
-  id: string;
-
   @Column("text", { nullable: false })
   title: string;
 
@@ -81,10 +51,4 @@ export class Job extends ExtendedBaseEntity {
 
   @ManyToOne(() => User, (user) => user.jobs, { nullable: false })
   user: User;
-
-  @CreateDateColumn({ name: "created_at" })
-  created_at: Date;
-
-  @UpdateDateColumn({ name: "updated_at" })
-  updated_at: Date;
 }
