+++ src/routes/job.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { JobController } from "../controllers/jobController";
-import { authMiddleware } from "../middleware";
+import { authMiddleware, validateData } from "../middleware";
+import { createJobSchema } from "../validationschema/job";
 
 const jobRoute = Router();
 
@@ -8,6 +9,7 @@ const jobController = new JobController();
 
 jobRoute.post(
   "/jobs",
+  validateData({ body: createJobSchema }),
   authMiddleware,
   jobController.createJob.bind(jobController),
 );
