+++ src/utils/index.ts
@@ -26,7 +26,7 @@ export const generateNumericOTP = (length: number): string => {
 };
 
 export const generateToken = (payload: Record<string, unknown>) => {
-  return jwt.sign(payload, "helloworld", {
+  return jwt.sign(payload, config.TOKEN_SECRET, {
     expiresIn: "1d",
   });
 };
